more work on entry models

+++ lib/src/main/scala/Libor.scala
@@ -7,7 +7,7 @@ import lib.syntax.*
 class Libor[T: DateLike](
     val currency: Currency,
     val tenor: Tenor,
-    val spotLag: Long,
+    val spotLag: Int,
     val dayCounter: DayCounter,
     val calendar: Calendar[T],
     val resetCurve: YieldCurve[T],
